+++ src/main/kotlin/com/github/jensim/megamanipulator/actions/ProcessOperator.kt
@@ -12,7 +12,7 @@ class ProcessOperator(
 ) {
 
     fun runCommandAsync(workingDir: File, command: List<String>): Deferred<ApplyOutput> {
-        val project = projectOperator.project!!
+        val project = projectOperator.project
         val tempOutput = File.createTempFile("mega-manipulator-apply-out", "txt")
         val tempErrOutput = File.createTempFile("mega-manipulator-apply-err", "txt")
         val proc = ProcessBuilder(command)