+++ src/main/kotlin/com/github/jensim/megamanipulator/ui/Extensions.kt
@@ -4,7 +4,7 @@ import com.intellij.openapi.project.Project
 import java.io.File
 
 fun File.trimProjectPath(project: Project): String {
-    val projectPath: String? = project?.basePath
+    val projectPath: String? = project.basePath
     return if (projectPath != null && this.path.startsWith(projectPath)) {
         this.path.drop(projectPath.length).let {
             if (it.startsWith('/')) {