+++ src/test/kotlin/com/github/jensim/megamanipulator/actions/search/sourcegraph/SourcegraphSearchClientTest.kt
@@ -7,23 +7,30 @@ import com.github.jensim.megamanipulator.settings.CodeHostSettings.GitHubSetting
 import com.github.jensim.megamanipulator.settings.ForkSetting.PLAIN_BRANCH
 import com.github.jensim.megamanipulator.settings.MegaManipulatorSettings
 import com.github.jensim.megamanipulator.settings.SearchHostSettings.SourceGraphSettings
+import com.github.jensim.megamanipulator.settings.SerializationHolder
 import com.github.jensim.megamanipulator.settings.SettingsFileOperator
+import com.github.jensim.megamanipulator.test.EnvHelper
+import com.github.jensim.megamanipulator.test.EnvHelper.EnvProperty.GITHUB_USERNAME
+import com.github.jensim.megamanipulator.test.EnvHelper.EnvProperty.SRC_COM_ACCESS_TOKEN
 import com.github.jensim.megamanipulator.test.TestPasswordOperator
 import io.mockk.every
 import io.mockk.mockk
 import kotlinx.coroutines.runBlocking
+import org.hamcrest.MatcherAssert.assertThat
+import org.hamcrest.Matchers.equalTo
 import org.junit.jupiter.api.Assertions.assertEquals
 import org.junit.jupiter.api.Test
 
 internal class SourcegraphSearchClientTest {
 
-    private val password = System.getenv("SRC_COM_ACCESS_TOKEN")
+    private val envHelper = EnvHelper()
+    private val password = envHelper.resolve(SRC_COM_ACCESS_TOKEN)
     private val codeHostName = "github.com"
     private val sourceGraphSettings = SourceGraphSettings(
         baseUrl = "https://sourcegraph.com",
         codeHostSettings = mapOf(
             codeHostName to GitHubSettings(
-                username = "jensim",
+                username = envHelper.resolve(GITHUB_USERNAME),
                 forkSetting = PLAIN_BRANCH,
             )
         )
@@ -37,16 +44,18 @@ internal class SourcegraphSearchClientTest {
     private val settingsMock: SettingsFileOperator = mockk {
         every { readSettings() } returns settings
     }
-    private val passwordsOperator = TestPasswordOperator(mapOf("token" to "https://sourcegraph.com" to password))
+    private val passwordsOperator = TestPasswordOperator(mapOf("token" to sourceGraphSettings.baseUrl to password))
     private val notificationsMock: NotificationsOperator = mockk()
     private val clientProvider = HttpClientProvider(
         settingsFileOperator = settingsMock,
         passwordsOperator = passwordsOperator,
         notificationsOperator = notificationsMock
     )
+    private val json = SerializationHolder().readableJson
     private val sourcegraphSearchClient = SourcegraphSearchClient(
         httpClientProvider = clientProvider,
-        notificationsOperator = notificationsMock
+        notificationsOperator = notificationsMock,
+        json = this.json,
     )
 
     @Test
@@ -68,5 +77,14 @@ internal class SourcegraphSearchClientTest {
                 )
             )
         )
+    }
+
+    @Test
+    internal fun `validate token test`() {
+        val result = runBlocking {
+            sourcegraphSearchClient.validateToken(searchHostName, sourceGraphSettings)
+        }
+
+        assertThat(result, equalTo("200:OK"))
     }
 }